Remove deprecated rules and add kit recipes
Remove the following rules:
- no-unnecessary-use-memo
- no-unnecessary-use-callback
- no-unused-state
- prefer-destructuring-assignment
- prefer-namespace-import (react-x and react-dom)

Add new @eslint-react/kit recipes:
- prefer-namespace-import recipe
- noDirectAccessProps in custom-rules-of-props

Update configs, documentation, and rule relations table accordingly.
