+++ apps/website/content/docs/recipes/index.mdx
@@ -73,10 +73,12 @@ export default defineConfig(
 
 ## Available Recipes
 
-| Recipe                                                           | Rules | Description                                                                                                                                 |
-| :--------------------------------------------------------------- | :---: | :------------------------------------------------------------------------------------------------------------------------------------------ |
-| [Component Hook Factories](./component-hook-factories)           |   1   | Disallows components or hooks defined inside other functions (factory pattern).                                                             |
-| [Custom Rules Of Props](./custom-rules-of-props)                 |   3   | Validates JSX props. Includes checks for duplicate props, mixing controlled and uncontrolled props, and explicit spread props.              |
-| [Custom Rules Of State](./custom-rules-of-state)                 |   1   | Validates state usage. Prefer the updater function form in useState setters.                                                                |
-| [Function Component Definition](./function-component-definition) |   1   | Enforces arrow function style for function component definitions.                                                                           |
-| [No Circular Effect](./no-circular-effect)                       |   1   | Detects circular dependencies between useEffect hooks. Prevents infinite update loops caused by effects that set state they also depend on. |
+| Recipe                                                           | Rules | Description                                                                                                                                         |
+| :--------------------------------------------------------------- | :---: | :-------------------------------------------------------------------------------------------------------------------------------------------------- |
+| [Component Hook Factories](/docs/recipes/component-hook-factories)           |   1   | Disallows components or hooks defined inside other functions (factory pattern).                                                                     |
+| [Custom Rules Of Props](/docs/recipes/custom-rules-of-props)                 |   4   | Validates JSX props. Includes checks for duplicate props, mixing controlled and uncontrolled props, explicit spread props, and direct props access. |
+| [Custom Rules Of State](/docs/recipes/custom-rules-of-state)                 |   1   | Validates state usage. Prefer the updater function form in useState setters.                                                                        |
+| [Destructure Component Props](/docs/recipes/destructure-component-props)     |   1   | Enforces destructuring assignment for component props instead of member expression access.                                                          |
+| [Function Component Definition](/docs/recipes/function-component-definition) |   1   | Enforces arrow function style for function component definitions.                                                                                   |
+| [No Circular Effect](/docs/recipes/no-circular-effect)                       |   1   | Detects circular dependencies between useEffect hooks. Prevents infinite update loops caused by effects that set state they also depend on.         |
+| [Prefer Namespace Import](/docs/recipes/prefer-namespace-import)             |   1   | Enforces importing React via a namespace import instead of default import.                                                                          |
