+++ apps/website/content/docs/rules/index.mdx
@@ -57,8 +57,6 @@ full: true
 | [`no-set-state-in-component-did-mount`](no-set-state-in-component-did-mount)       | 1️⃣ 1️⃣ |           | Disallows calling `this.setState` in `componentDidMount` outside functions such as callbacks                                                                                |          |
 | [`no-set-state-in-component-did-update`](no-set-state-in-component-did-update)     | 1️⃣ 1️⃣ |           | Disallows calling `this.setState` in `componentDidUpdate` outside functions such as callbacks                                                                               |          |
 | [`no-set-state-in-component-will-update`](no-set-state-in-component-will-update)   | 1️⃣ 1️⃣ |           | Disallows calling `this.setState` in `componentWillUpdate` outside functions such as callbacks                                                                              |          |
-| [`no-unnecessary-use-callback`](no-unnecessary-use-callback)                       | 0️⃣ 1️⃣ |    `🧪`   | Disallows unnecessary usage of `useCallback`                                                                                                                                |          |
-| [`no-unnecessary-use-memo`](no-unnecessary-use-memo)                               | 0️⃣ 1️⃣ |    `🧪`   | Disallows unnecessary usage of `useMemo`                                                                                                                                    |          |
 | [`no-unnecessary-use-prefix`](no-unnecessary-use-prefix)                           | 1️⃣ 1️⃣ |           | Enforces that a function with the `use` prefix uses at least one Hook inside it                                                                                             |          |
 | [`no-unsafe-component-will-mount`](no-unsafe-component-will-mount)                 | 1️⃣ 1️⃣ |           | Warns about the use of `UNSAFE_componentWillMount` in class components                                                                                                      |          |
 | [`no-unsafe-component-will-receive-props`](no-unsafe-component-will-receive-props) | 1️⃣ 1️⃣ |           | Warns about the use of `UNSAFE_componentWillReceiveProps` in class components                                                                                               |          |
@@ -67,10 +65,7 @@ full: true
 | [`no-unstable-default-props`](no-unstable-default-props)                           | 0️⃣ 1️⃣ |    `⚙️`   | Prevents using referential-type values as default props in object destructuring                                                                                             |          |
 | [`no-unused-class-component-members`](no-unused-class-component-members)           | 1️⃣ 1️⃣ |           | Warns about unused class component methods and properties                                                                                                                   |          |
 | [`no-unused-props`](no-unused-props)                                               | 0️⃣ 0️⃣ | `💭` `🧪` | Warns about component props that are defined but never used                                                                                                                 |          |
-| [`no-unused-state`](no-unused-state)                                               | 0️⃣ 1️⃣ |           | Warns about unused class component state                                                                                                                                    |          |
 | [`no-use-context`](no-use-context)                                                 | 1️⃣ 1️⃣ |    `🔄`   | Replaces usage of `useContext` with `use`                                                                                                                                   | >=19.0.0 |
-| [`prefer-destructuring-assignment`](prefer-destructuring-assignment)               | 0️⃣ 1️⃣ |           | Enforces destructuring assignment for component props and context                                                                                                           |          |
-| [`prefer-namespace-import`](prefer-namespace-import)                               | 0️⃣ 0️⃣ |    `🔧`   | Enforces importing React via a namespace import                                                                                                                             |          |
 | [`purity`](purity)                                                                 | 1️⃣ 1️⃣ |           | Validates that components and hooks are pure by checking that they do not call known-impure functions during render                                                         |          |
 | [`refs`](refs)                                                                     | 0️⃣ 0️⃣ |    `🧪`   | Validates correct usage of refs by checking that `ref.current` is not read or written during render                                                                         |          |
 | [`rules-of-hooks`](rules-of-hooks)                                                 | 2️⃣ 2️⃣ | `⚙️` `🔧` | Enforces the Rules of Hooks                                                                                                                                                 |          |
@@ -131,7 +126,6 @@ full: true
 | [`no-unsafe-target-blank`](dom-no-unsafe-target-blank)                                         | 0️⃣ 1️⃣ |    `🔧`   | Disallows `target="_blank"` without `rel="noreferrer noopener"`                               |             |
 | [`no-use-form-state`](dom-no-use-form-state)                                                   | 2️⃣ 2️⃣ |    `🔄`   | Replaces usage of `useFormState` with `useActionState`                                        |   >=19.0.0  |
 | [`no-void-elements-with-children`](dom-no-void-elements-with-children)                         | 2️⃣ 2️⃣ |           | Disallows `children` in void DOM elements                                                     |             |
-| [`prefer-namespace-import`](dom-prefer-namespace-import)                                       | 0️⃣ 0️⃣ |    `🔧`   | Enforces importing React DOM via a namespace import                                           |             |
 
 ## Web API Rules
