+++ apps/website/content/docs/rules/meta.json
@@ -35,8 +35,6 @@
     "no-set-state-in-component-did-mount",
     "no-set-state-in-component-did-update",
     "no-set-state-in-component-will-update",
-    "no-unnecessary-use-callback",
-    "no-unnecessary-use-memo",
     "no-unnecessary-use-prefix",
     "no-unsafe-component-will-mount",
     "no-unsafe-component-will-receive-props",
@@ -45,10 +43,7 @@
     "no-unstable-default-props",
     "no-unused-class-component-members",
     "no-unused-props",
-    "no-unused-state",
     "no-use-context",
-    "prefer-destructuring-assignment",
-    "prefer-namespace-import",
     "purity",
     "refs",
     "rules-of-hooks",
@@ -85,7 +80,6 @@
     "dom-no-unsafe-target-blank",
     "dom-no-use-form-state",
     "dom-no-void-elements-with-children",
-    "dom-prefer-namespace-import",
     "---Web API Rules---",
     "web-api-no-leaked-event-listener",
     "web-api-no-leaked-interval",
