--- packages/plugins/eslint-plugin-react-dom/src/rules/prefer-namespace-import/prefer-namespace-import.mdx
@@ -1,50 +1,0 @@
----
-title: prefer-namespace-import
-description: Enforces importing React DOM via a namespace import.
----
-
-**Full Name in `eslint-plugin-react-dom`**
-
-```plain copy
-react-dom/prefer-namespace-import
-```
-
-**Full Name in [`@eslint-react/eslint-plugin`](https://npmx.dev/package/@eslint-react/eslint-plugin/v/latest)**
-
-```plain copy
-@eslint-react/dom-prefer-namespace-import
-```
-
-**Features**
-
-`🔧`
-
-**Presets**
-
--
-
-## Rule Details
-
-Namespace imports (`import * as ReactDOM from 'react-dom'`) make it clearer which APIs are being used from the React DOM package.
-
-## Common Violations
-
-### Invalid
-
-```tsx
-import ReactDOM from "react-dom/client";
-
-import type ReactDOM from "react-dom/client";
-```
-
-### Valid
-
-```tsx
-import * as ReactDOM from "react-dom/client";
-import type * as ReactDOM from "react-dom/client";
-```
-
-## Resources
-
-- [Rule Source](https://github.com/Rel1cx/eslint-react/tree/main/packages/plugins/eslint-plugin-react-dom/src/rules/prefer-namespace-import/prefer-namespace-import.ts)
-- [Test Source](https://github.com/Rel1cx/eslint-react/tree/main/packages/plugins/eslint-plugin-react-dom/src/rules/prefer-namespace-import/prefer-namespace-import.spec.ts)
