+++ packages/plugins/eslint-plugin-react-rsc/src/utils/create-rule.ts
@@ -1,8 +1,7 @@
-import { WEBSITE_URL } from "@eslint-react/shared";
 import { ESLintUtils } from "@typescript-eslint/utils";
 
 function getDocsUrl(ruleName: string) {
-  return `${WEBSITE_URL}/docs/rules/${ruleName}`;
+  return `https://eslint-react.xyz/docs/rules/${ruleName}`;
 }
 
 export const createRule = ESLintUtils.RuleCreator(getDocsUrl);
