+++ packages/plugins/eslint-plugin-react-x/src/configs/disable-experimental.ts
@@ -9,8 +9,6 @@ export const rules: Record<string, RuleConfig> = {
   "react-x/no-implicit-key": "off",
   "react-x/no-implicit-ref": "off",
   "react-x/no-misused-capture-owner-stack": "off",
-  "react-x/no-unnecessary-use-callback": "off",
-  "react-x/no-unnecessary-use-memo": "off",
   "react-x/no-unused-props": "off",
   "react-x/refs": "off",
   "react-x/set-state-in-render": "off",
