+++ packages/plugins/eslint-plugin-react-x/src/configs/strict.ts
@@ -7,12 +7,8 @@ export const rules = {
   ...recommended.rules,
   "react-x/no-class-component": "error",
   "react-x/no-misused-capture-owner-stack": "error",
-  "react-x/no-unnecessary-use-callback": "warn",
-  "react-x/no-unnecessary-use-memo": "warn",
   "react-x/no-unstable-context-value": "warn",
   "react-x/no-unstable-default-props": "warn",
-  "react-x/no-unused-state": "warn",
-  "react-x/prefer-destructuring-assignment": "warn",
 } as const satisfies Record<string, RuleConfig>;
 
 export const plugins = {
