--- packages/plugins/eslint-plugin-react-x/src/rules/no-unnecessary-use-callback/no-unnecessary-use-callback.mdx
@@ -1,101 +1,0 @@
----
-title: no-unnecessary-use-callback
-description: Disallows unnecessary usage of 'useCallback'.
----
-
-<Callout type="warning">
-  This rule is experimental and may change in the future or be removed. It is not recommended for use in production code at this time.
-</Callout>
-
-**Full Name in [`eslint-plugin-react-x`](https://npmx.dev/package/eslint-plugin-react-x/v/latest)**
-
-```plain copy
-react-x/no-unnecessary-use-callback
-```
-
-**Full Name in [`@eslint-react/eslint-plugin`](https://npmx.dev/package/@eslint-react/eslint-plugin/v/latest)**
-
-```plain copy
-@eslint-react/no-unnecessary-use-callback
-```
-
-**Features**
-
-`🧪`
-
-**Presets**
-
-`strict`
-`strict-typescript`
-`strict-type-checked`
-
-## Rule Details
-
-The React Hook `useCallback` with an empty dependency array, as in the examples, is unnecessary. The hook can be removed and its value can be created in the component body or hoisted to the outer scope of the component.
-If the calculated function is only used inside one `useEffect`, the calculation can be moved inside the `useEffect` function.
-
-## Common Violations
-
-### Invalid
-
-```tsx
-import React, { useCallback } from "react";
-
-function MyComponent() {
-  const onClick = useCallback(() => {
-    console.log("clicked");
-  }, []);
-
-  return <button type="button" onClick={onClick} />;
-}
-```
-
-```tsx
-import { Button, MantineTheme } from "@mantine/core";
-import React, { useCallback, useEffect } from "react";
-
-function MyComponent({items}: {items: string[]}) {
-  const updateTest = useCallback(() => {console.log(items.length)}, [items]);
-
-  useEffect(() => {
-    updateTest();
-  }, [updateTest]);
-
-  return <div>Hello World</div>;
-}
-```
-
-### Valid
-
-```tsx
-function onClick() {
-  console.log("clicked");
-}
-
-function MyComponent() {
-  return <button type="button" onClick={onClick} />;
-}
-```
-
-```tsx
-import { Button, MantineTheme } from "@mantine/core";
-import React, { useEffect } from "react";
-
-function MyComponent({items}: {items: string[]}) {
-  useEffect(() => {
-    console.log(items.length);
-  }, [items]);
-
-  return <div>Hello World</div>;
-}
-```
-
-## Resources
-
-- [Rule Source](https://github.com/Rel1cx/eslint-react/tree/main/packages/plugins/eslint-plugin-react-x/src/rules/no-unnecessary-use-callback/no-unnecessary-use-callback.ts)
-- [Test Source](https://github.com/Rel1cx/eslint-react/tree/main/packages/plugins/eslint-plugin-react-x/src/rules/no-unnecessary-use-callback/no-unnecessary-use-callback.spec.ts)
-
-## Further Reading
-
-- [React Docs: `useCallback`](https://react.dev/reference/react/useCallback)
-- [React Docs: `useEffect`](https://react.dev/reference/react/useEffect)
