--- packages/plugins/eslint-plugin-react-x/src/rules/no-unnecessary-use-memo/no-unnecessary-use-memo.mdx
@@ -1,111 +1,0 @@
----
-title: no-unnecessary-use-memo
-description: Disallows unnecessary usage of 'useMemo'.
----
-
-<Callout type="warning">
-  This rule is experimental and may change in the future or be removed. It is not recommended for use in production code at this time.
-</Callout>
-
-**Full Name in [`eslint-plugin-react-x`](https://npmx.dev/package/eslint-plugin-react-x/v/latest)**
-
-```plain copy
-react-x/no-unnecessary-use-memo
-```
-
-**Full Name in [`@eslint-react/eslint-plugin`](https://npmx.dev/package/@eslint-react/eslint-plugin/v/latest)**
-
-```plain copy
-@eslint-react/no-unnecessary-use-memo
-```
-
-**Features**
-
-`🧪`
-
-**Presets**
-
-`strict`
-`strict-typescript`
-`strict-type-checked`
-
-## Rule Details
-
-The React Hook `useMemo` with an empty dependencies array, as in the examples, is unnecessary. The hook can be removed, and its value can be calculated in the component body or hoisted to the outer scope of the component. If the calculated variable is used only inside one `useEffect`, the calculation can be moved inside the `useEffect` function.
-
-## Common Violations
-
-### Invalid
-
-```tsx
-import { Button, MantineTheme } from "@mantine/core";
-import React, { useMemo } from "react";
-
-function MyComponent() {
-  const style = useMemo(
-    (theme: MantineTheme) => ({
-      input: {
-        fontFamily: theme.fontFamilyMonospace,
-      },
-    }),
-    [],
-  );
-  return <Button sx={style} />;
-}
-```
-
-```tsx
-import { Button, MantineTheme } from "@mantine/core";
-import React, { useMemo, useEffect } from "react";
-
-function MyComponent({ someNumbers }: { someNumbers: number[] }) {
-  const calculatedNumber = useMemo(
-    () => someNumbers.reduce((prev, next) => prev + next, 0),
-    [someNumbers],
-  );
-
-  useEffect(() => {
-    console.log(calculatedNumber);
-  }, [calculatedNumber]);
-  return <div>Hello World!</div>;
-}
-```
-
-### Valid
-
-```tsx
-import { Button, MantineTheme } from "@mantine/core";
-
-const style = (theme: MantineTheme) => ({
-  input: {
-    fontFamily: theme.fontFamilyMonospace,
-  },
-});
-
-function MyComponent() {
-  return <Button sx={style} />;
-}
-```
-
-```tsx
-import { Button, MantineTheme } from "@mantine/core";
-import React, { useEffect } from "react";
-
-function MyComponent({ someNumbers }: { someNumbers: number[] }) {
-  useEffect(() => {
-    const calculatedNumber = someNumbers.reduce((prev, next) => prev + next, 0);
-    console.log(calculatedNumber);
-  }, [someNumbers]);
-  return <div>Hello World!</div>;
-}
-```
-
-## Resources
-
-- [Rule Source](https://github.com/Rel1cx/eslint-react/tree/main/packages/plugins/eslint-plugin-react-x/src/rules/no-unnecessary-use-memo/no-unnecessary-use-memo.ts)
-- [Test Source](https://github.com/Rel1cx/eslint-react/tree/main/packages/plugins/eslint-plugin-react-x/src/rules/no-unnecessary-use-memo/no-unnecessary-use-memo.spec.ts)
-
-## Further Reading
-
-- [React Docs: `useMemo`](https://react.dev/reference/react/useMemo)
-- [React Docs: `useEffect`](https://react.dev/reference/react/useEffect)
