--- packages/plugins/eslint-plugin-react-x/src/rules/no-unnecessary-use-memo/no-unnecessary-use-memo.ts
@@ -1,165 +1,0 @@
-import * as ast from "@eslint-react/ast";
-import * as core from "@eslint-react/core";
-import { type RuleContext, type RuleFeature, defineRuleListener, report } from "@eslint-react/shared";
-import { resolve } from "@eslint-react/var";
-import { identity } from "@local/eff";
-import type { Scope } from "@typescript-eslint/scope-manager";
-import { AST_NODE_TYPES as AST, type TSESTree } from "@typescript-eslint/types";
-import { isIdentifier, isVariableDeclarator } from "@typescript-eslint/utils/ast-utils";
-import type { ReportDescriptor, SourceCode } from "@typescript-eslint/utils/ts-eslint";
-import { match } from "ts-pattern";
-
-import { createRule } from "../../utils";
-
-export const RULE_NAME = "no-unnecessary-use-memo";
-
-export const RULE_FEATURES = [
-  "EXP",
-] as const satisfies RuleFeature[];
-
-export type MessageID = "default" | "noUnnecessaryUseMemoInsideUseEffect";
-
-export default createRule<[], MessageID>({
-  meta: {
-    type: "suggestion",
-    docs: {
-      description: "Disallows unnecessary usage of 'useMemo'.",
-    },
-    messages: {
-      default: "An 'useMemo' with empty deps and no references to the component scope may be unnecessary.",
-      noUnnecessaryUseMemoInsideUseEffect:
-        "{{name}} is only used inside 1 useEffect, which may be unnecessary. You can move the computation into useEffect directly and merge the dependency arrays.",
-    },
-    schema: [],
-  },
-  name: RULE_NAME,
-  create,
-  defaultOptions: [],
-});
-
-export function create(context: RuleContext<MessageID, []>) {
-  // Fast path: skip if `useMemo` is not present in the file
-  if (!context.sourceCode.text.includes("useMemo")) return {};
-  return defineRuleListener(
-    {
-      VariableDeclarator(node) {
-        const { id, init } = node;
-        if (id.type !== AST.Identifier || init?.type !== AST.CallExpression || !core.isUseMemoCall(context, init)) {
-          return;
-        }
-        const [mem, ...rest] = context.sourceCode.getDeclaredVariables(node);
-        // Skip non-standard `useMemo()` usages to prevent false positives
-        if (mem == null || rest.length > 0) return;
-
-        const checkForUsageInsideUseEffectReport = checkForUsageInsideUseEffect(context.sourceCode, init);
-        const scope = context.sourceCode.getScope(init);
-        const component = scope.block;
-        if (!ast.isFunction(component)) {
-          return;
-        }
-        const [arg0, arg1] = init.arguments;
-        if (arg0 == null || arg1 == null) {
-          return;
-        }
-        const hasCallInArg0 = ast.isFunction(arg0)
-          && [...ast.getNestedCallExpressions(arg0.body), ...ast.getNestedNewExpressions(arg0.body)].length > 0;
-
-        if (hasCallInArg0) {
-          report(context)(checkForUsageInsideUseEffectReport);
-          return;
-        }
-
-        const hasEmptyDeps = match(arg1)
-          .with({ type: AST.ArrayExpression }, (n) => n.elements.length === 0)
-          .with({ type: AST.Identifier }, (n) => {
-            const initNode = resolve(context, n);
-            if (initNode?.type !== AST.ArrayExpression) {
-              return false;
-            }
-            return initNode.elements.length === 0;
-          })
-          .otherwise(() => false);
-
-        if (!hasEmptyDeps) {
-          report(context)(checkForUsageInsideUseEffectReport);
-          return;
-        }
-        const arg0Node = match(arg0)
-          .with({ type: AST.ArrowFunctionExpression }, (n) => {
-            if (n.body.type === AST.ArrowFunctionExpression) {
-              return n.body;
-            }
-            return n;
-          })
-          .with({ type: AST.FunctionExpression }, identity)
-          .with({ type: AST.Identifier }, (n) => {
-            const initNode = resolve(context, n);
-            if (initNode?.type !== AST.ArrowFunctionExpression && initNode?.type !== AST.FunctionExpression) {
-              return null;
-            }
-            return initNode;
-          })
-          .otherwise(() => null);
-        if (arg0Node == null) return;
-
-        function getChildScopes(scope: Scope): Scope[] {
-          return scope.childScopes.reduce((acc, child) => [...acc, ...getChildScopes(child)], [scope]);
-        }
-        const arg0NodeScope = context.sourceCode.getScope(arg0Node);
-        const arg0NodeReferences = getChildScopes(arg0NodeScope).flatMap((x) => x.references);
-        const isReferencedToComponentScope = arg0NodeReferences.some((x) => x.resolved?.scope.block === component);
-
-        if (!isReferencedToComponentScope) {
-          context.report({
-            messageId: "default",
-            node,
-          });
-          return;
-        }
-        report(context)(checkForUsageInsideUseEffectReport);
-      },
-    },
-  );
-}
-
-function checkForUsageInsideUseEffect(
-  sourceCode: Readonly<SourceCode>,
-  node: TSESTree.CallExpression,
-): ReportDescriptor<MessageID> | null {
-  if (!/use\w*Effect/u.test(sourceCode.text)) return null;
-
-  if (!isVariableDeclarator(node.parent)) {
-    return null;
-  }
-
-  if (!isIdentifier(node.parent.id)) {
-    return null;
-  }
-
-  const references = sourceCode.getDeclaredVariables(node.parent)[0]?.references ?? [];
-  const usages = references.filter((ref) => ref.init !== true);
-  // Skip if there are no usages of the memoized value, the no-unused-vars rule will catch it
-  if (usages.length === 0) {
-    return null;
-  }
-  const effectSet = new Set<TSESTree.Node>();
-
-  for (const usage of usages) {
-    const effect = ast.findParent(usage.identifier, core.isUseEffectLikeCall);
-
-    if (effect == null) {
-      return null;
-    }
-
-    effectSet.add(effect);
-    if (effectSet.size > 1) {
-      return null;
-    }
-  }
-
-  return {
-    data: { name: node.parent.id.name },
-    messageId: "noUnnecessaryUseMemoInsideUseEffect",
-    node,
-  };
-}
