--- packages/plugins/eslint-plugin-react-x/src/rules/no-unused-state/no-unused-state.mdx
@@ -1,70 +1,0 @@
----
-title: no-unused-state
-description: Warns about unused class component state.
----
-
-**Full Name in [`eslint-plugin-react-x`](https://npmx.dev/package/eslint-plugin-react-x/v/latest)**
-
-```plain copy
-react-x/no-unused-state
-```
-
-**Full Name in [`@eslint-react/eslint-plugin`](https://npmx.dev/package/@eslint-react/eslint-plugin/v/latest)**
-
-```plain copy
-@eslint-react/no-unused-state
-```
-
-**Presets**
-
-`strict`
-`strict-typescript`
-`strict-type-checked`
-
-## Rule Details
-
-This rule warns about state properties in class components that are defined but never used. Unused state can lead to unnecessary re-renders and confusion about the component's actual data requirements.
-
-## Common Violations
-
-### Invalid
-
-```tsx
-import React from "react";
-
-class MyComponent extends React.Component {
-  // Unused
-  state = {
-    foo: 1,
-  };
-
-  render() {
-    return null;
-  }
-}
-```
-
-### Valid
-
-```tsx
-import React from "react";
-
-class MyComponent extends React.Component {
-  state = {
-    foo: 1,
-  };
-
-  render() {
-    return this.state.foo;
-  }
-}
-```
-
-## Resources
-
-- [Rule Source](https://github.com/Rel1cx/eslint-react/tree/main/packages/plugins/eslint-plugin-react-x/src/rules/no-unused-state/no-unused-state.ts)
-- [Test Source](https://github.com/Rel1cx/eslint-react/tree/main/packages/plugins/eslint-plugin-react-x/src/rules/no-unused-state/no-unused-state.spec.ts)
-
-## Further Reading
-
-- [Legacy React APIs: `Component`](https://react.dev/reference/react/Component)
