--- packages/plugins/eslint-plugin-react-x/src/rules/prefer-namespace-import/prefer-namespace-import.mdx
@@ -1,57 +1,0 @@
----
-title: prefer-namespace-import
-description: Enforces importing React via a namespace import.
----
-
-**Full Name in [`eslint-plugin-react-x`](https://npmx.dev/package/eslint-plugin-react-x/v/latest)**
-
-```plain copy
-react-x/prefer-namespace-import
-```
-
-**Full Name in [`@eslint-react/eslint-plugin`](https://npmx.dev/package/@eslint-react/eslint-plugin/v/latest)**
-
-```plain copy
-@eslint-react/prefer-namespace-import
-```
-
-**Features**
-
-`🔧`
-
-**Presets**
-
--
-
-## Rule Details
-
-## Common Violations
-
-### Invalid
-
-```tsx
-import React from "react";
-
-import type React from "react";
-
-import React, { useState } from "react";
-
-import type React, { useState } from "react";
-```
-
-### Valid
-
-```tsx
-import * as React from "react";
-
-import type * as React from "react";
-
-import { useState } from "react";
-
-import type { useState } from "react";
-```
-
-## Resources
-
-- [Rule Source](https://github.com/Rel1cx/eslint-react/tree/main/packages/plugins/eslint-plugin-react-x/src/rules/prefer-namespace-import/prefer-namespace-import.ts)
-- [Test Source](https://github.com/Rel1cx/eslint-react/tree/main/packages/plugins/eslint-plugin-react-x/src/rules/prefer-namespace-import/prefer-namespace-import.spec.ts)
