--- packages/plugins/eslint-plugin-react-x/src/rules/prefer-namespace-import/prefer-namespace-import.spec.ts
@@ -1,102 +1,0 @@
-import { AST_NODE_TYPES as AST } from "@typescript-eslint/types";
-
-import { ruleTester } from "../../../../../../test";
-import rule, { RULE_NAME } from "./prefer-namespace-import";
-
-ruleTester.run(RULE_NAME, rule, {
-  invalid: [
-    {
-      code: `import React from 'react';`,
-      errors: [{ type: AST.ImportDeclaration, messageId: "default" }],
-      output: `import * as React from 'react';`,
-    },
-    {
-      code: `import REACT from 'react';`,
-      errors: [{ type: AST.ImportDeclaration, messageId: "default" }],
-      output: `import * as REACT from 'react';`,
-    },
-    {
-      code: `import type React from 'react';`,
-      errors: [{ type: AST.ImportDeclaration, messageId: "default" }],
-      output: `import type * as React from 'react';`,
-    },
-    {
-      code: `import React, {useState} from 'react';`,
-      errors: [{ type: AST.ImportDefaultSpecifier, messageId: "default" }],
-      output: `import * as React from 'react';\nimport {useState} from 'react';`,
-    },
-    {
-      code: `import React, {useState, useReducer} from 'react';`,
-      errors: [{ type: AST.ImportDefaultSpecifier, messageId: "default" }],
-      output: `import * as React from 'react';\nimport {useState, useReducer} from 'react';`,
-    },
-    {
-      code: `import REACT, {useState} from 'react';`,
-      errors: [{ type: AST.ImportDefaultSpecifier, messageId: "default" }],
-      output: `import * as REACT from 'react';\nimport {useState} from 'react';`,
-    },
-    {
-      code: `import React from "react";`,
-      errors: [{ type: AST.ImportDeclaration, messageId: "default" }],
-      output: `import * as React from "react";`,
-    },
-    {
-      code: `import REACT from "react";`,
-      errors: [{ type: AST.ImportDeclaration, messageId: "default" }],
-      output: `import * as REACT from "react";`,
-    },
-    {
-      code: `import type React from "react";`,
-      errors: [{ type: AST.ImportDeclaration, messageId: "default" }],
-      output: `import type * as React from "react";`,
-    },
-    {
-      code: `import React, {useState} from "react";`,
-      errors: [{ type: AST.ImportDefaultSpecifier, messageId: "default" }],
-      output: `import * as React from "react";\nimport {useState} from "react";`,
-    },
-    {
-      code: `import React, {useState, useReducer} from "react";`,
-      errors: [{ type: AST.ImportDefaultSpecifier, messageId: "default" }],
-      output: `import * as React from "react";\nimport {useState, useReducer} from "react";`,
-    },
-    {
-      code: `import REACT, {useState} from "react";`,
-      errors: [{ type: AST.ImportDefaultSpecifier, messageId: "default" }],
-      output: `import * as REACT from "react";\nimport {useState} from "react";`,
-    },
-    {
-      code: `import React from 'react'`,
-      errors: [{ type: AST.ImportDeclaration, messageId: "default" }],
-      output: `import * as React from 'react'`,
-    },
-  ],
-  valid: [
-    {
-      code: `import * as React from 'react';`,
-    },
-    {
-      code: `import {useState} from 'react';`,
-    },
-    {
-      code: `import {} from 'react';`,
-    },
-    {
-      code: `import * as React from "react";`,
-    },
-    {
-      code: `import {useState} from "react";`,
-    },
-    {
-      code: `import {} from "react";`,
-    },
-    {
-      code: `import * as React from "@pika/react";`,
-      settings: {
-        "react-x": {
-          importSource: "@pika/react",
-        },
-      },
-    },
-  ],
-});
