+++ packages/plugins/eslint-plugin/src/configs/all.ts
@@ -35,8 +35,6 @@ export const rules = {
   "@eslint-react/no-set-state-in-component-did-mount": "warn",
   "@eslint-react/no-set-state-in-component-did-update": "warn",
   "@eslint-react/no-set-state-in-component-will-update": "warn",
-  "@eslint-react/no-unnecessary-use-callback": "warn",
-  "@eslint-react/no-unnecessary-use-memo": "warn",
   "@eslint-react/no-unnecessary-use-prefix": "warn",
   "@eslint-react/no-unsafe-component-will-mount": "warn",
   "@eslint-react/no-unsafe-component-will-receive-props": "warn",
@@ -44,10 +42,7 @@ export const rules = {
   "@eslint-react/no-unstable-context-value": "warn",
   "@eslint-react/no-unstable-default-props": "warn",
   "@eslint-react/no-unused-class-component-members": "warn",
-  "@eslint-react/no-unused-state": "warn",
   "@eslint-react/no-use-context": "warn",
-  "@eslint-react/prefer-destructuring-assignment": "warn",
-  "@eslint-react/prefer-namespace-import": "warn",
   "@eslint-react/purity": "warn",
   "@eslint-react/refs": "error",
   "@eslint-react/rules-of-hooks": "error",
@@ -83,8 +78,6 @@ export const rules = {
   "@eslint-react/dom-no-unsafe-target-blank": "warn",
   "@eslint-react/dom-no-use-form-state": "warn",
   "@eslint-react/dom-no-void-elements-with-children": "error",
-  "@eslint-react/dom-prefer-namespace-import": "warn",
-
   "@eslint-react/web-api-no-leaked-event-listener": "warn",
   "@eslint-react/web-api-no-leaked-interval": "warn",
   "@eslint-react/web-api-no-leaked-resize-observer": "warn",
