+++ packages/plugins/eslint-plugin/src/configs/disable-experimental.ts
@@ -9,8 +9,6 @@ export const rules: Record<string, RuleConfig> = {
   "@eslint-react/no-implicit-key": "off",
   "@eslint-react/no-implicit-ref": "off",
   "@eslint-react/no-misused-capture-owner-stack": "off",
-  "@eslint-react/no-unnecessary-use-callback": "off",
-  "@eslint-react/no-unnecessary-use-memo": "off",
   "@eslint-react/no-unused-props": "off",
   "@eslint-react/refs": "off",
   "@eslint-react/set-state-in-render": "off",
