+++ packages/plugins/eslint-plugin/src/configs/strict.ts
@@ -9,12 +9,8 @@ export const rules = {
 
   "@eslint-react/no-class-component": "error",
   "@eslint-react/no-misused-capture-owner-stack": "error",
-  "@eslint-react/no-unnecessary-use-callback": "warn",
-  "@eslint-react/no-unnecessary-use-memo": "warn",
   "@eslint-react/no-unstable-context-value": "warn",
   "@eslint-react/no-unstable-default-props": "warn",
-  "@eslint-react/no-unused-state": "warn",
-  "@eslint-react/prefer-destructuring-assignment": "warn",
 
   "@eslint-react/jsx-no-children-prop": "error",
   "@eslint-react/jsx-no-useless-fragment": "warn",
