--- packages/shared/docs/variables/GITHUB_URL.md
@@ -1,9 +1,0 @@
-[@eslint-react/shared](../README.md) / GITHUB\_URL
-
-# Variable: GITHUB\_URL
-
-```ts
-const GITHUB_URL: "https://github.com/Rel1cx/eslint-react" = "https://github.com/Rel1cx/eslint-react";
-```
-
-The GitHub repository for this project.
