--- packages/shared/docs/variables/IMPURE_CTORS.md
@@ -1,9 +1,0 @@
-[@eslint-react/shared](../README.md) / IMPURE\_CTORS
-
-# Variable: IMPURE\_CTORS
-
-```ts
-const IMPURE_CTORS: ReadonlySet<string>;
-```
-
-Known impure global constructors used with `new`
