--- packages/shared/docs/variables/IMPURE_FUNCS.md
@@ -1,9 +1,0 @@
-[@eslint-react/shared](../README.md) / IMPURE\_FUNCS
-
-# Variable: IMPURE\_FUNCS
-
-```ts
-const IMPURE_FUNCS: ReadonlyMap<string, ReadonlySet<string>>;
-```
-
-Known impure functions
