--- packages/shared/docs/variables/NPM_SCOPE.md
@@ -1,9 +1,0 @@
-[@eslint-react/shared](../README.md) / NPM\_SCOPE
-
-# Variable: NPM\_SCOPE
-
-```ts
-const NPM_SCOPE: "@eslint-react" = "@eslint-react";
-```
-
-The NPM scope for this project.
