--- packages/shared/docs/variables/PURE_CTORS.md
@@ -1,9 +1,0 @@
-[@eslint-react/shared](../README.md) / PURE\_CTORS
-
-# Variable: PURE\_CTORS
-
-```ts
-const PURE_CTORS: ReadonlySet<string>;
-```
-
-Known pure global constructors used with `new`
