--- packages/shared/docs/variables/PURE_FUNCS.md
@@ -1,9 +1,0 @@
-[@eslint-react/shared](../README.md) / PURE\_FUNCS
-
-# Variable: PURE\_FUNCS
-
-```ts
-const PURE_FUNCS: ReadonlyMap<string, ReadonlySet<string>>;
-```
-
-Known pure functions
