--- packages/shared/docs/variables/WEBSITE_URL.md
@@ -1,9 +1,0 @@
-[@eslint-react/shared](../README.md) / WEBSITE\_URL
-
-# Variable: WEBSITE\_URL
-
-```ts
-const WEBSITE_URL: "https://eslint-react.xyz" = "https://eslint-react.xyz";
-```
-
-The URL to the project's website.
