+++ packages/shared/src/index.ts
@@ -1,7 +1,5 @@
-export * from "./constants";
 export * from "./define-rule-listener";
 export * from "./id-generator";
-export * from "./react-version";
 export * from "./regexp";
 export * from "./report";
 export * from "./settings";
